Add new trap scripts and animations for Fire and Spikedball traps
