+++ Assets/Animations/Traps/Fire.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 431936e2dd0df2d44b170ddbc21251f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
