+++ Assets/Animations/Traps/Fire/AC_TrapFire.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f10b1be5c42f03c48a5e2e28883ab9d4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
