+++ Assets/Animations/Traps/Fire/AC_TrapFireButton.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d97a7c0b326ebd418914e42537197d1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
