+++ Assets/Animations/Traps/Fire/ButtonPressed.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 08318464b2f6ee24eb63334287693ab7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
