+++ Assets/Animations/Traps/Fire/buttonNotPressed.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9375525c138e3e47a1a7800f115c220
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
