+++ Assets/Animations/Traps/Fire/fire off.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e9ad0c18902b0242aabbc4cea0b519e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
