+++ Assets/Animations/Traps/Fire/fire on.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fd10969cc64ffdb478cb241d9b33ed81
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
