+++ Assets/Prefabs/Traps/Trap_Fire.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7878f2a3d5954f3489896bf8beb703e9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
