+++ Assets/Prefabs/Traps/Trap_FireButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c9e1252b819bce442bfdad0b1c88cfa5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
