+++ Assets/Scripts/Trap_FireButton.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e31a6fede098ba544a81c44ab499fa65
