+++ Assets/Scripts/Traps.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 481348c84e0626a4b8d3488d95e879f7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
